+++ lib/src/ui/table_advanced.dart
@@ -18,6 +18,12 @@ class TableAdvancedColumnHeader {
   /// will be called. This can be useful i.e. to sort elements in the table.
   final VoidCallback? onTap;
 
+  /// If you want to sort columns, implement this method.
+  ///
+  /// Note that you have to manage sorting externally and update the table UI accordingly
+  /// using the `sortedAsc` parameter.
+  final void Function(bool sortedAsc)? onSortTapped;
+
   /// Configuration for column headers of [TableAdvanced].
   ///
   /// The `child` is the [Widget] shown as column header. Use `flex` to specify
@@ -27,6 +33,7 @@ class TableAdvancedColumnHeader {
     required this.child,
     this.flex = 1,
     this.onTap,
+    this.onSortTapped,
   });
 }
 
@@ -81,6 +88,15 @@ class TableAdvanced<T> extends StatefulWidget {
   /// You can define the spacing between rows. Defaults to 12.
   final double rowSpacing;
 
+  /// Use this builder to customize the pagination controls when [TableMode] is paginationPage.
+  ///
+  /// If `null`, a default pagination will be shown
+  final Widget Function(TableAdvancedController<T> controller)?
+      paginationBuilder;
+
+  /// A [Widget] to show when there is no data to display. Defaults to empty space.
+  final Widget? emptyState;
+
   /// An easy to use table with responsive layout and pagination.
   ///
   /// Use the `controller` to manipulate table properties such as content and pagination.
@@ -92,6 +108,8 @@ class TableAdvanced<T> extends StatefulWidget {
     required this.rowBuilder,
     required this.controller,
     this.rowSpacing = 12,
+    this.paginationBuilder,
+    this.emptyState,
   }) : super(key: key);
 
   @override
@@ -104,6 +122,26 @@ class _TableAdvancedState<T> extends State<TableAdvanced<T>> {
 
   late ScrollController _headerController;
   late ScrollController _bodyController;
+
+  int? _sortedIndex;
+  bool _sortAsc = true;
+
+  void _changeSort(int index) {
+    if (_sortedIndex == index) {
+      _sortAsc = !_sortAsc;
+    } else {
+      _sortedIndex = index;
+      _sortAsc = true;
+    }
+    widget.columnHeaders[index].onSortTapped?.call(_sortAsc);
+  }
+
+  bool? _isSortedAsc(int index) {
+    if (_sortedIndex != index) {
+      return null;
+    }
+    return _sortAsc;
+  }
 
   @override
   void initState() {
@@ -136,10 +174,10 @@ class _TableAdvancedState<T> extends State<TableAdvanced<T>> {
       child: Builder(
         builder: (context) {
           if (context
-              .read<TableAdvancedController<T>>()
+              .watch<TableAdvancedController<T>>()
               .dataItemsToShow
               .isEmpty) {
-            return const SizedBox.shrink();
+            return widget.emptyState ?? const SizedBox.shrink();
           }
 
           var showScrollBar =
@@ -183,7 +221,7 @@ class _TableAdvancedState<T> extends State<TableAdvanced<T>> {
               ),
               if (context.read<TableAdvancedController<T>>().mode ==
                   TableMode.paginationPage)
-                TableAdvancedPagination(controller: widget.controller),
+                _paginationWidget(),
             ],
           ),
         );
@@ -206,8 +244,8 @@ class _TableAdvancedState<T> extends State<TableAdvanced<T>> {
                     widget.controller.dataItemsToShow.length,
                 onChanged: (checked) {
                   widget.controller.checkItems(
-                      widget.controller.dataItemsToShow,
-                      checkAll: true,
+                    widget.controller.dataItemsToShow,
+                    checkAll: true,
                   );
                 },
               ),
@@ -222,9 +260,29 @@ class _TableAdvancedState<T> extends State<TableAdvanced<T>> {
                     child: Material(
                       child: InkWell(
                         onTap: widget.columnHeaders[index].onTap,
-                        child: Padding(
-                          padding: const EdgeInsets.symmetric(horizontal: 10),
-                          child: widget.columnHeaders[index].child,
+                        child: Row(
+                          children: [
+                            Padding(
+                              padding:
+                                  const EdgeInsets.symmetric(horizontal: 10),
+                              child: widget.columnHeaders[index].child,
+                            ),
+                            if (widget.columnHeaders[index].onSortTapped !=
+                                null)
+                              IconButton(
+                                icon: Opacity(
+                                    opacity:
+                                        _isSortedAsc(index) == null ? 0.5 : 1,
+                                    child: _isSortedAsc(index) ?? false
+                                        ? const Icon(
+                                            Icons.keyboard_arrow_up_rounded)
+                                        : const Icon(
+                                            Icons.keyboard_arrow_down_rounded)),
+                                onPressed: () {
+                                  _changeSort(index);
+                                },
+                              ),
+                          ],
                         ),
                       ),
                     ),
@@ -240,8 +298,7 @@ class _TableAdvancedState<T> extends State<TableAdvanced<T>> {
 
   Widget _tableRows() {
     return Builder(builder: (context) {
-      var dataRows =
-          context.watch<TableAdvancedController<T>>().dataItemsToShow;
+      var dataRows = context.read<TableAdvancedController<T>>().dataItemsToShow;
 
       return ListView.separated(
         itemCount: dataRows.length,
@@ -287,8 +344,9 @@ class _TableAdvancedState<T> extends State<TableAdvanced<T>> {
                               onChanged: (item.disabled ?? false)
                                   ? null
                                   : (checked) {
-                                      widget.controller.checkItems([
-                                        dataRows[index],
+                                      widget.controller.checkItems(
+                                        [
+                                          dataRows[index],
                                         ],
                                       );
                                     },
@@ -338,6 +396,16 @@ class _TableAdvancedState<T> extends State<TableAdvanced<T>> {
           });
         },
       );
+    });
+  }
+
+  Widget _paginationWidget() {
+    if (widget.paginationBuilder == null) {
+      return TableAdvancedPagination(controller: widget.controller);
+    }
+    return Builder(builder: (context) {
+      return widget.paginationBuilder!
+          .call(context.watch<TableAdvancedController<T>>());
     });
   }
 }