+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+* Added functionality to sort columns
+* Added pagination controls customization
+
 ## 0.0.2
 Added docs and improved Readme.